Do Not Save Or Load Search String
We are loading the saved search string into the input field but we are
not loading the search results when the page loads. We are instead
showing the main page. This may cause confusion. It is better not to
save search string altogether as it does not have any significant benefit.